Changed Authenticate + IsUserDuplicate()
