+++ Karma/BusinessLogic/DB/User.cs
@@ -12,20 +12,20 @@ namespace BusinessLogic
     /// </summary>
     public class User
     {
-        public User (string _Name, string _LastName, string _UserName, string _Password)
+        
+        public string Name { get; set; }
+        public string LastName { get; set; }
+        public string UserName { get; set; }
+        public string Password { get; set; }
+        public User(string _Name, string _LastName, string _UserName, string _Password)
         {
             Name = _Name;
             LastName = _LastName;
             UserName = _UserName;
             Password = _Password;
         }
-        public string Name { get; set; }
-        public string LastName { get; set; }
-        public string UserName { get; set; }
-        public string Password { get; set; }
-
         //Stuff like House Adress, Phone number (they can also be optional if people don't want to give their data)
-        
+
         //Token might be useful for authentification stuff maybe? 
         //string Token;
